backend/matching-engine/image-recommendation/requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform
google-cloud-storage
spacy
tensorflow
tensorflow_hub
tensorflow_text
pandas
db-dtypes